requirements.txt
fastapi
uvicorn
openai
pymongo
motor
python-dotenv
google-generativeai
gtts